old-tests/submission/Microsoft/selection/collapseToStart.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title id="desc">HTML5 Selection: Collapse the selection with collapseToStart()</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="common.js"></script>
        <script type="text/javascript">
            function RunTest()
            {
                var selection = window.getSelection();
                var p1 = document.getElementById("p1");

                var range = document.createRange();
                range.selectNode(p1);
                selection.addRange(range);
                selection.collapseToStart();

                range.collapse(true);

                checkSelectionAttributes(range.startContainer, range.startOffset, range.endContainer, range.endOffset, true, 1);
                assert_equals(selection.toString(), "")
            }
        </script>
    </head>
    <body onload="test(RunTest);">
        <p id="p1">Collapse the selection with collapseToStart()</p>
    </body>
</html>
